The port takes a text path the original does not
The garbled message line on the copy-protection screen has a diagnosis. Asked
about the driver's fast glyph blit from the copy-protection snapshot, the
verifier answered "captured 0 calls" - **the original never reaches it on that
screen**. The port does; that is what the abort said when it was still a stub.

So the fault is not inside the glyph blit at all. The port is choosing the fast
path where the original chooses the slow one, and the two draw differently
because the fast path is a driver entry and the slow path is `draw_char` with
its five style bits.

Which of the three conditions differs is not yet established, but reading them
again found that they are not all the same kind of comparison: the style at
0x3892 is `jle`, **signed**; the clip flag at 0x3893 is sign-extended with
`cbw` before being tested; and the font marker at 0x6176 is `jbe`, unsigned.
The transcription had all three unsigned. They agree on every value this game
is known to use and disagree on a style byte of 0x80 or more, so this is not
the cause of what is on the screen - but it is a difference, and it is fixed
rather than left because the next screen may be the one that has it.

Also added the glyph blit to the verifier with register arguments and its
overlay offset, which is what made the "0 calls" answer possible at all.

+++ reconstruct/seg1c25.c
@@ -5132,7 +5132,16 @@ void draw_string_body(uint16_t str, int16_t x, int16_t y)
     if (str == 0)
         return;
 
-    if (DG8(0x3892) <= 1 && DG8(0x3893) == 0 && DG8(0x6176) <= 1) {
+    /*
+     * The three tests are not all the same kind. The style at 0x3892 is
+     * compared with `jle` - **signed**, so a style byte with bit 7 set passes
+     * it - the clip flag at 0x3893 is sign-extended with `cbw` before being
+     * tested against zero, and the font marker at 0x6176 is compared with
+     * `jbe`, unsigned. Written as three unsigned tests they would agree on
+     * every value this game uses and disagree on a style of 0x80 or more.
+     */
+    if ((int8_t)DG8(0x3892) <= 1 && (int8_t)DG8(0x3893) == 0
+        && DG8(0x6176) <= 1) {
         /*
          * The fast path: each character goes straight to the driver, and only
          * a character **wider than 8 pixels** falls back to `draw_char` -
